requirements.txt
alembic==1.4.0
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.5
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.1.0
Flask-Migrate==2.5.2
flask-restplus==0.13.0
flask-restplus-sqlalchemy==0.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
PyJWT==2.0.1
pylint==2.4.4
pyrsistent==0.15.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
sentry-sdk==0.14.2
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
zipp==3.0.0
